+++ src/app.module.ts
@@ -6,6 +6,8 @@ import { CustomerModule } from './customer/customer.module';
 import { Customer } from './customer/entities/customer.entity';
 import { StoreModule } from './store/store.module';
 import { Store } from './store/entities/store.entity';
+import { ProductModule } from './product/product.module';
+import { Product } from './product/entities/product.entity';
 
 @Module({
   imports: [TypeOrmModule.forRoot({
@@ -14,9 +16,9 @@ import { Store } from './store/entities/store.entity';
     port: 5432,
     password: '1234',
     username: 'postgres',
-    entities: [Customer, Store],
+    entities: [Customer, Store, Product],
     database: 'commerce'
-    }), CustomerModule, StoreModule],
+    }), CustomerModule, StoreModule, ProductModule],
   controllers: [AppController],
   providers: [AppService],
 })